requirements.txt
quart==0.19.4
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.0
Pillow==10.1.0
requests==2.31.0
google-generativeai==0.3.2
cohere==4.37
PyJWT==2.8.0
bcrypt==4.1.2 